requirements.txt
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
drf-yasg
django-filter